requirements.txt
openai==1.2.3
requests==2.31.0

# for bot
python-telegram-bot==20.6
python-dotenv==1.0.0

# for bot db connection
psycopg2-binary==2.9.9

Django==4.2.7
asgiref==3.7.2
